+++ HISTORY.md
@@ -5,6 +5,7 @@
   objects using the Dependency Injection pattern.
 * Add JsonCodecableWithServicesTestTrait, a PHPUnit test trait to verify
   that a class correctly implements JsonCodecableWithServicesTrait.
+* Add Hint::hasModifier() helper function.
 
 ## 5.0.0 (2026-02-12)
 * Allow a Hint to be the first argument to Hint::build().